rename config models
