+++ packages/unraid-api-plugin-connect/src/job/timeout-checker.job.ts
@@ -3,7 +3,7 @@ import { SchedulerRegistry } from '@nestjs/schedule';
 
 import { isDefined } from 'class-validator';
 
-import { MinigraphStatus } from '../model/config.entity.js';
+import { MinigraphStatus } from '../model/connect-config.model.js';
 import { ONE_MINUTE_MS, THREE_MINUTES_MS } from '../helper/generic-consts.js';
 import { DynamicRemoteAccessService } from '../service/dynamic-remote-access.service.js';
 import { MothershipConnectionService } from '../service/connection.service.js';
