+++ packages/unraid-api-plugin-connect/src/model/cloud.model.ts
@@ -1,6 +1,6 @@
 import { Field, Int, ObjectType } from '@nestjs/graphql';
 
-import { MinigraphStatus } from './my-servers-config.js';
+import { MinigraphStatus } from './my-servers-config.model.js';
 
 @ObjectType()
 export class ApiKeyResponse {
