+++ packages/unraid-api-plugin-connect/src/service/connect-config.service.ts
@@ -2,7 +2,7 @@ import { Injectable, Logger } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { OnEvent } from '@nestjs/event-emitter';
 
-import { ConfigType, emptyMyServersConfig, MyServersConfig } from '../model/config.entity.js';
+import { ConfigType, emptyMyServersConfig, MyServersConfig } from '../model/connect-config.model.js';
 import { EVENTS } from '../helper/nest-tokens.js';
 
 @Injectable()
